+++ prisma/migrations/20240407073037_added_currency_to_order/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Order" ADD COLUMN "currency" TEXT;